Switched to an SVP-based MenuModule system  * Basically, getting data from the ui is easier now.  * Note that I still need to fix the layout of the MenuModule.

+++ src/stringvaluepair.cpp
@@ -1,6 +1,6 @@
 #include "stringvaluepair.h"
 
-SVP::SVP(QString s, int v)
+SVP::SVP(QString s, double v)
 {
     string = s;
     value = v;
